+++ src/c/container.c
@@ -424,16 +424,6 @@ void iot_container_delete_component (iot_container_t * cont, const char * name)
   pthread_rwlock_unlock (&cont->lock);
 }
 
-static iot_data_t * iot_container_component_data_locked (const iot_component_t * comp)
-{
-  iot_data_t * data = iot_data_alloc_map (IOT_DATA_STRING);
-  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.name), iot_data_alloc_string (comp->name, IOT_DATA_REF));
-  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.type), iot_data_alloc_string (comp->factory->type, IOT_DATA_REF));
-  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.state), iot_data_alloc_string (iot_component_state_name (comp->state), IOT_DATA_REF));
-  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.config), iot_data_add_ref (comp->config));
-  return data;
-}
-
 iot_data_t * iot_container_list_components (iot_container_t * cont)
 {
   assert (cont);
@@ -443,14 +433,13 @@ iot_data_t * iot_container_list_components (iot_container_t * cont)
   iot_data_list_iter (cont->components, &iter);
   while (iot_data_list_iter_next (&iter))
   {
-    iot_data_t * data = iot_container_component_data_locked (iot_data_list_iter_pointer_value (&iter));
-    iot_data_list_tail_push (list, data);
+    iot_data_list_tail_push (list, iot_component_read (iot_data_list_iter_pointer_value (&iter)));
   }
   pthread_rwlock_unlock (&cont->lock);
   return list;
 }
 
-iot_data_t * iot_container_get_component_data (iot_container_t * cont, const char * name)
+iot_data_t * iot_container_component_read (iot_container_t * cont, const char * name)
 {
   assert (cont);
   iot_data_t * data = NULL;
@@ -458,7 +447,7 @@ iot_data_t * iot_container_get_component_data (iot_container_t * cont, const cha
   {
     pthread_rwlock_rdlock (&cont->lock);
     const iot_component_t * comp = iot_container_find_component_locked (cont, name);
-    if (comp) data = iot_container_component_data_locked (comp);
+    if (comp) data = iot_component_read (comp);
     pthread_rwlock_unlock (&cont->lock);
   }
   return data;